+++ realtime.h
@@ -1,3 +1,3 @@
-int   RealSearch(long maxnodes);
-short MoveValue(MAZE *maze, MOVE *move, long effort);
+int   RealSearch(int32_t maxnodes);
+short MoveValue(MAZE *maze, MOVE *move, int32_t effort);
